+++ StomtiOS/LikeWishView.h
@@ -1,0 +1,14 @@
+//
+//  LikeWishView.h
+//  StomtiOS
+//
+//  Created by Leonardo Cascianelli on 02/03/16.
+//  Copyright © 2016 Leonardo Cascianelli. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+#import "Stomt.h"
+
+@interface LikeWishView : UIView
+- (void)setupWithIdentifier:(kSTObjectQualifier)likeOrWish;
+@end