+++ StomtiOS/TargetView.h
@@ -1,0 +1,18 @@
+//
+//  TargetView.h
+//  StomtiOS
+//
+//  Created by Leonardo Cascianelli on 03/03/16.
+//  Copyright © 2016 Leonardo Cascianelli. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@class STTarget;
+@class STImageView;
+
+@interface TargetView : UIView
+@property (nonatomic,strong) STImageView* targetImage;
+@property (nonatomic,strong) UILabel* targetName;
+- (void)setupWithTarget:(STTarget*)target;
+@end