+++ StomtiOS/block_declarations.h
@@ -21,4 +21,5 @@ typedef void (^ StomtCreationBlock)(NSError* error, STObject* stomt);
 typedef void (^ ImageUploadBlock)(NSError* error, STImage* image);
 typedef void (^ FeedRequestBlock)(NSError* error, STFeed* feed);
 typedef void (^ TargetRequestBlock)(NSError* error, STTarget* target);
+typedef void (^ UserAvailabilityBlock)(NSError* error, NSNumber* available);
 #endif